+++ cmd/payout.go
@@ -12,7 +12,7 @@ import (
 )
 
 var (
-	secret             string
+	privateKey         string
 	totalReward        string
 	rawLoadbalancerUrl string
 
@@ -42,31 +42,33 @@ var payoutCmd = &cobra.Command{
 
 func init() {
 	payoutCmd.Flags().StringVar(
-		&secret,
-		"secret",
+		&privateKey,
+		"private-key",
 		"",
-		"[REQUIRED] loadbalancer wallet secret",
+		"[REQUIRED] loadbalancer wallet private key",
 	)
-	_ = payoutCmd.MarkFlagRequired("secret")
 	payoutCmd.Flags().StringVar(
 		&totalReward,
-		"reward",
+		"payout-reward",
 		"",
 		"[REQUIRED] total reward pool in Planck",
 	)
-	_ = payoutCmd.MarkFlagRequired("total-reward")
 	payoutCmd.Flags().StringVar(
 		&rawLoadbalancerUrl,
 		"load-balancer-url",
 		"http://localhost:80",
 		"[OPTIONAL] url on which loadbalancer is listening")
+
+	_ = startCmd.MarkFlagRequired("private-key")
+	_ = startCmd.MarkFlagRequired("payout-reward")
+
 	RootCmd.AddCommand(payoutCmd)
 }
 
 func payoutCommand(_ *cobra.Command, _ []string) {
 	DisplayBanner()
 	fmt.Println("Payout script running...")
-	transactions, err := script.ExecutePayout(secret, totalRewardAsFloat64, loadbalancerURL)
+	transactions, err := script.ExecutePayout(privateKey, totalRewardAsFloat64, loadbalancerURL)
 	if transactions != nil {
 		// display even if only part of transactions executed
 		ui.DisplayTransactionsStatus(transactions)