+++ internal/payout/payout.go
@@ -8,29 +8,14 @@ import (
 	"time"
 )
 
-type NodePayoutDetails struct {
-	Stats         models.NodeStatsDetails `json:"stats"`
-	PayoutAddress string                  `json:"payout_address"`
-}
-
 func GetStatsForPayout(
 	repos repositories.Repos,
 	intervalEnd time.Time,
 	recordPayout bool,
-) (map[string]models.NodeStatsDetails, map[string]NodePayoutDetails, error) {
-
+) (map[string]models.NodeStatsDetails, error) {
 	statistics, err := stats.CalculateStatisticsFromLastPayout(repos, intervalEnd)
 	if err != nil {
-		return nil, nil, err
-	}
-
-	payoutStatistics := make(map[string]NodePayoutDetails, len(statistics))
-	for nodeId, statsDetails := range statistics {
-		node, _ := repos.NodeRepo.FindByID(nodeId)
-		payoutStatistics[nodeId] = NodePayoutDetails{
-			Stats:         statsDetails,
-			PayoutAddress: node.PayoutAddress,
-		}
+		return nil, err
 	}
 
 	if recordPayout {
@@ -40,9 +25,9 @@ func GetStatsForPayout(
 		})
 		if err != nil {
 			log.Errorf("Unable to save payout information To database, because of: %v", err)
-			return nil, nil, err
+			return nil, err
 		}
 	}
 
-	return statistics, payoutStatistics, nil
+	return statistics, nil
 }