+++ internal/rpc/rpc_test.go
@@ -153,11 +153,11 @@ func TestSendRequestToNode(t *testing.T) {
 			if tt.args.url == "valid" && tt.portValid {
 				serverURL, _ := url.Parse(server.URL)
 				port, _ := strconv.Atoi(serverURL.Port())
-				poolerMock.On("GetPort", mock.Anything).Once().Return(port, nil)
+				poolerMock.On("GetHTTPPort", mock.Anything).Once().Return(port, nil)
 			} else if tt.args.url == "invalid" && tt.portValid {
-				poolerMock.On("GetPort", mock.Anything).Once().Return(1331313, nil)
+				poolerMock.On("GetHTTPPort", mock.Anything).Once().Return(1331313, nil)
 			} else {
-				poolerMock.On("GetPort", mock.Anything).Once().Return(0, fmt.Errorf("ERROR"))
+				poolerMock.On("GetHTTPPort", mock.Anything).Once().Return(0, fmt.Errorf("ERROR"))
 			}
 
 			mux.HandleFunc("/", tt.handleFunc)