+++ internal/script/payout.go
@@ -9,34 +9,30 @@ import (
 	"net/url"
 
 	"github.com/NodeFactoryIo/vedran/internal/controllers"
-	"github.com/NodeFactoryIo/vedran/internal/models"
 	"github.com/NodeFactoryIo/vedran/internal/payout"
+	log "github.com/sirupsen/logrus"
+	"net/http"
+	"net/url"
+	"strconv"
 )
 
-var statsEndpoint, _ = url.Parse("/api/v1/stats")
-
 func ExecutePayout(secret string, totalReward float64, loadbalancerUrl *url.URL) ([]*payout.TransactionDetails, error) {
-	stats, err := fetchStatsFromEndpoint(loadbalancerUrl.ResolveReference(statsEndpoint), secret)
+	log.Infof("New payout started with total reward: %s", strconv.FormatFloat(totalReward, 'f', 0, 64))
+	response, err := fetchStatsFromEndpoint(statsEndpoint(loadbalancerUrl))
 	if err != nil {
 		return nil, fmt.Errorf("unable to fetch stats from loadbalancer, %v", err)
 	}
 
-	// calculate distribution
-	nodeStatsDetails := make(map[string]models.NodeStatsDetails, len(stats.Stats))
-	for nodeId, nodeStats := range stats.Stats {
-		nodeStatsDetails[nodeId] = nodeStats.Stats
-	}
 	distributionByNode := payout.CalculatePayoutDistributionByNode(
-		nodeStatsDetails,
+		response.Stats,
 		totalReward,
-		float64(stats.Fee),
+		float64(response.Fee),
 	)
 
 	return payout.ExecuteAllPayoutTransactions(
 		distributionByNode,
-		stats.Stats,
 		secret,
-		loadbalancerUrl.String(),
+		wsEndpoint(loadbalancerUrl).String(),
 	)
 }