+++ internal/ui/table.go
@@ -10,9 +10,9 @@ func DisplayTransactionsStatus(transactions []*payout.TransactionDetails) {
 	table := uitable.New()
 	table.MaxColWidth = 80
 	table.Wrap = true
-	table.AddRow("ID (node)", "To", "Amount", "Status")
+	table.AddRow("To (Node)", "Amount", "Status")
 	for _, tx := range transactions {
-		table.AddRow(tx.NodeId, tx.To, tx.Amount.String(), tx.Status)
+		table.AddRow(tx.To, tx.Amount.String(), tx.Status)
 	}
 	fmt.Println(table)
 }